requirements.txt
streamlit~=1.39.0
PyPDF2~=3.0.1
python-dotenv~=1.0.0
cohere~=5.11.0
langchain~=0.3.2